backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
pandas==2.2.3
openpyxl==3.1.5
google-generativeai==0.8.3
slowapi==0.1.9
pydantic-settings==2.6.1
python-dotenv==1.0.1
aiofiles==24.1.0
httpx==0.27.2
ruff==0.7.4
